:sparkles: onClick 위치 변경
